README
Quality Evaluation of Modern Code Reviews Through Intelligent Biometric Program Comprehension/Supplements
This supplement includes the metadata and the datasets that would enable other researchers to reproduce the experiment. The files are: a)The experiment protocol b)The metadata of the code sections under review performance c) the EEG data of the subjects (anonymous) d) The Physiological data of the subjects (anonymous), c) the screening test and the C programming test, d) the codes with the bugs locations.